# docs/openapi-review.yaml
openapi: 3.0.3
info:
  title: DevSmith Review Service API
  description: |
    AI-powered code analysis service with 5 reading modes for effective code comprehension.
    
    **Reading Modes:**
    - **Preview**: Quick 2-3 minute structural overview
    - **Skim**: 5-7 minute surface-level scan of architecture
    - **Scan**: 3-5 minute targeted search for specific patterns
    - **Detailed**: 10-15 minute deep line-by-line analysis
    - **Critical**: 15-20 minute quality evaluation and issue detection
    
    **Architecture:**
    - Circuit breaker protection (5 failures → open, 60s timeout)
    - Graceful shutdown (30s timeout for in-flight requests)
    - Health checks every 30s
    - OpenTelemetry tracing to Jaeger
  version: 1.0.0
  contact:
    name: DevSmith Platform Team
    url: https://github.com/mikejsmith1985/devsmith-modular-platform
servers:
  - url: http://localhost:8081
    description: Local development
  - url: http://localhost:3000
    description: Local via nginx gateway
paths:
...
  /api/review/modes/preview:
    post:
      summary: Analyze code in Preview mode
      description: Quick structural overview (2-3 minutes)
      operationId: analyzePreview
      tags:
        - Analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalysisRequest'
      responses:
        '200':
          description: Analysis result (HTML for HTMX)
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Invalid request
          content:
            text/html:
              schema:
                type: string
        '503':
          description: Service unavailable (Ollama down or circuit open)
          content:
            text/html:
              schema:
                type: string
...
components:
  schemas:
...
    AnalysisRequest:
      type: object
      required:
        - pasted_code
      properties:
        pasted_code:
          type: string
          description: Code to analyze
          example: |
            package main
            
            func main() {
              println("Hello world")
            }
        model:
          type: string
          description: Ollama model to use
          default: mistral:7b-instruct
          enum:
            - mistral:7b-instruct
            - codellama:13b
            - llama2:13b
            - deepseek-coder:6.7b
            - deepseek-coder-v2:16b
...
  securitySchemes: {}
...